+++ languages/de/default.php
@@ -52,3 +52,4 @@ $GLOBALS['TL_LANG']['MSC']['downloadSize'] = '%s';
 $GLOBALS['TL_LANG']['MSC']['downloadSizePdf'] = '%s (%s)';
 
 $GLOBALS['TL_LANG']['MSC']['downloadPdfItem'] = 'PDF herunterladen';
+$GLOBALS['TL_LANG']['MSC']['downloadItem'] = 'herunterladen';